apps/agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.5.0",
    "google-cloud-aiplatform[adk,agent-engines]==1.94.0",
    "google-cloud-logging>=3.12.1",
    "nest-asyncio>=1.6.0",
    "pydub>=0.25.1",
]
